Move external user-based tests into folder
Like `/internal` the Cypress `/external` tests were also broken up into sub folders.

In this first change, we move those equivalent to the internal user tests into an external `/user` sub folder.
